+++ src/ChecksForValidTypes.php
@@ -6,7 +6,7 @@ use InvalidArgumentException;
 
 trait ChecksForValidTypes
 {
-    /** @var list<class-string> */
+    /** @var list<class-string|non-empty-string> */
     protected static $allowedTypes = [];
 
     /**
@@ -26,14 +26,16 @@ trait ChecksForValidTypes
      */
     protected function assertValidType($item): void
     {
+        $itemType = get_debug_type($item);
+
         foreach (static::$allowedTypes as $allowedType) {
-            if ($item instanceof $allowedType) {
+            if (($itemType === $allowedType) || ($item instanceof $allowedType)) {
                 return;
             }
         }
 
         throw new InvalidArgumentException(sprintf(
-            'A %s collection only accepts objects of the following type(s): %s.',
+            'A %s collection only accepts items of the following type(s): %s.',
             get_class($this), implode(', ', static::$allowedTypes)
         ));
     }