+++ tests/LazyMixedTypeCollection.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gamez\Illuminate\Support\Tests;
+
+use Gamez\Illuminate\Support\LazyTypedCollection;
+
+final class LazyMixedTypeCollection extends LazyTypedCollection
+{
+    protected static $allowedTypes = ['int', 'string', ArrayableItem::class];
+}